Shortened test_make_debug_pipeline.py to two tests and merged with test_debug_pipeline.py to avoid redundancy and improve maintainability
