+++ tests/test_pipeline/test_debug_pipeline.py
@@ -11,9 +11,9 @@ from sklearn.multiclass import (
 )
 from sklearn.pipeline import FeatureUnion
 from sklearn.svm import LinearSVC
+from sklearn.preprocessing import StandardScaler
 
-from sklego.pipeline import DebugPipeline
-
+from sklego.pipeline import DebugPipeline, make_debug_pipeline
 
 IRIS = datasets.load_iris()
 
@@ -54,13 +54,29 @@ def custom_log_callback(output, execution_time, **kwargs):
 
 
 @pytest.fixture
-def steps():
+def named_steps():
     return [
         ("add_1", Adder(value=1)),
         ("add_10", Adder(value=10)),
         ("add_100", Adder(value=100)),
         ("add_1000", Adder(value=1000)),
     ]
+
+@pytest.fixture
+def nameless_steps():
+    return (
+        Adder(value=1),
+        Adder(value=10),
+        Adder(value=100),
+        Adder(value=1000)
+    )
+
+@pytest.fixture
+def repeated_steps():
+    return (
+        StandardScaler(),
+        StandardScaler()
+    )
 
 
 @pytest.mark.filterwarnings("ignore: The default of the `iid`")  # 0.22
@@ -77,16 +93,16 @@ def test_classifier_gridsearch(cls):
     assert best_C in Cs
 
 
-def test_no_logs_when_log_callback_is_None(caplog, steps):
-    pipe = DebugPipeline(steps, log_callback=None)
+def test_no_logs_when_log_callback_is_None(caplog, named_steps):
+    pipe = DebugPipeline(named_steps, log_callback=None)
     caplog.clear()
     with caplog.at_level(logging.INFO):
         pipe.fit(IRIS.data, IRIS.target)
     assert not caplog.text, f"Log should be empty: {caplog.text}"
 
 
-def test_output_shape_in_logs_when_log_callback_is_default(caplog, steps):
-    pipe = DebugPipeline(steps, log_callback="default")
+def test_output_shape_in_logs_when_log_callback_is_default(caplog, named_steps):
+    pipe = DebugPipeline(named_steps, log_callback="default")
     caplog.clear()
     with caplog.at_level(logging.INFO):
         pipe.fit(IRIS.data, IRIS.target)
@@ -98,8 +114,8 @@ def test_output_shape_in_logs_when_log_callback_is_default(caplog, steps):
     ), f'"{shape_str}" should be {len(pipe.steps) - 1} times in {caplog.text}'
 
 
-def test_time_in_logs_when_log_callback_is_default(caplog, steps):
-    pipe = DebugPipeline(steps, log_callback="default")
+def test_time_in_logs_when_log_callback_is_default(caplog, named_steps):
+    pipe = DebugPipeline(named_steps, log_callback="default")
     caplog.clear()
     with caplog.at_level(logging.INFO):
         pipe.fit(IRIS.data, IRIS.target)
@@ -110,8 +126,8 @@ def test_time_in_logs_when_log_callback_is_default(caplog, steps):
     ), f'"time" should be {len(pipe.steps) - 1} times in {caplog.text}'
 
 
-def test_step_name_in_logs_when_log_callback_is_default(caplog, steps):
-    pipe = DebugPipeline(steps, log_callback="default")
+def test_step_name_in_logs_when_log_callback_is_default(caplog, named_steps):
+    pipe = DebugPipeline(named_steps, log_callback="default")
     caplog.clear()
     with caplog.at_level(logging.INFO):
         pipe.fit(IRIS.data, IRIS.target)
@@ -123,8 +139,8 @@ def test_step_name_in_logs_when_log_callback_is_default(caplog, steps):
         ), f"{step} should be once in {caplog.text}"
 
 
-def test_nbytes_in_logs_when_log_callback_is_custom(caplog, steps):
-    pipe = DebugPipeline(steps, log_callback=custom_log_callback)
+def test_nbytes_in_logs_when_log_callback_is_custom(caplog, named_steps):
+    pipe = DebugPipeline(named_steps, log_callback=custom_log_callback)
     caplog.clear()
     with caplog.at_level(logging.INFO):
         pipe.fit(IRIS.data, IRIS.target)
@@ -135,9 +151,9 @@ def test_nbytes_in_logs_when_log_callback_is_custom(caplog, steps):
     ), f'"nbytes=" should be {len(pipe.steps) - 1} times in {caplog.text}'
 
 
-def test_feature_union(caplog, steps):
-    pipe_w_default_log_callback = DebugPipeline(steps, log_callback="default")
-    pipe_w_custom_log_callback = DebugPipeline(steps, log_callback=custom_log_callback)
+def test_feature_union(caplog, named_steps):
+    pipe_w_default_log_callback = DebugPipeline(named_steps, log_callback="default")
+    pipe_w_custom_log_callback = DebugPipeline(named_steps, log_callback=custom_log_callback)
 
     pipe_union = FeatureUnion(
         [
@@ -156,3 +172,21 @@ def test_feature_union(caplog, steps):
             assert (
                 caplog.text.count(str(step)) == 2
             ), f"{step} should be once in {caplog.text}"
+
+
+def test_different_name_for_repeated_step(nameless_steps):
+    ss_twice_pipeline = make_debug_pipeline(*nameless_steps)
+    assert ss_twice_pipeline.steps[0][0] != ss_twice_pipeline.steps[1][0]
+
+
+def test_nameless_step_name_in_logs_when_log_callback_is_default(caplog, nameless_steps):
+    pipe = make_debug_pipeline(*nameless_steps, log_callback="default")
+    caplog.clear()
+    with caplog.at_level(logging.INFO):
+        pipe.fit(IRIS.data, IRIS.target)
+    assert caplog.text, f"Log should be none empty: {caplog.text}"
+    for _, step in pipe.steps[:-1]:
+        assert str(step) in caplog.text, f"{step} should be in: {caplog.text}"
+        assert (
+            caplog.text.count(str(step)) == 1
+        ), f"{step} should be once in {caplog.text}"
